Rework truncate() logic
We _should_ be including leading sigils in the excerpt, then trimming them at render time. The alternative would require us to fetch the entire memo body eagerly for each item in a list view just to render correct formatting with leading sigils which throws away the benefits of calculating excerpt in the first place.