Update search to use cursor
In my playtesting, this significantly improves the perf of type-ahead
suggestions. I suspect this may have something to do with the way inputs
work in SwiftUI. Perhaps the binding-style closer access that goes all
the way up to store allows the input field to render text BEFORE the
search view is invalidated by the parent.