Do not refresh viewer if content is indentical (#969)
Fixes #940 

Introduce new `MemoViewer` action `.refreshAll` that actually does the
loading. `.appear` will only trigger loading if the address changes and
the `MemoViewerDetailView` listens for
`AppAction.succeedSyncSphereWithGateway` to trigger `.refreshAll`.

My thinking is that 3P content can only change after a sync, so that's
the only explicit trigger we need.