BlockEditor driven by shared ObservableStore (#1000)
Fixes #987. 

Depends on
https://github.com/subconsciousnetwork/ObservableStore/pull/43

This PR implements basic loading and saving of real data from database,
including blocks and related notes. I've kept the scope to *just* saving
and loading, and there is more work to be done wrt transcludes and
loading states.

## Approach

- New changes to ObservableStore allow creating custom Update types for
models. https://github.com/subconsciousnetwork/ObservableStore/pull/43
- Also introduces an `updates` publisher on stores
- Custom update type lets us include a change description along with
state changes, supplying extra information that the ViewController might
need to perform these changes.
- Root UIViewController subscribes to `updates` publisher and performs
updates to controller based on state transactions.
- ObservableStore instance is stored as a `@StateObject` on the
enclosing SwiftUIView and passed down as an `@ObservedObject` to the
UIViewControllerRepresentable.

The result is that we can share the same store between SwiftUI and UIKit
land, as well as the same environment and side effects management.