Fix multiplayer links (#1014)
Fixes https://github.com/subconsciousnetwork/subconscious/issues/1002

I've extracted all the logic for link mapping to one function, reworked
the approach and added tests and plenty of comments. The bug was due to
passing the wrong DID to an action at one of the view call sites so I've
tried to remove that footgun from the equation going forward.